+++ brian2/importexport/__init__.py
@@ -3,8 +3,11 @@ Package providing import/export support.
 '''
 
 from .importexport import *
-from .explicit import *
+from .dictlike import *
+
+
+__all__ = ['ImportExport']
 
 # Register the standard ImportExport methods
 for obj in [DictImportExport(), PandasImportExport()]:
-    ImportExport.register(obj.name, obj)
+    ImportExport.register(obj)