+++ brian2/importexport/dictlike.py
@@ -1,3 +1,7 @@
+'''
+Module providing `DictImportExport` and `PandasImportExport` (requiring a
+working installation of pandas).
+'''
 import numpy as np
 from importexport import ImportExport
 
@@ -38,15 +42,16 @@ class PandasImportExport(ImportExport):
 
     @staticmethod
     def export_data(group, variables, units=True, level=0):
-        # as pandas is not a default brian2 dependency we import it only in that namespace
+        # pandas is not a default brian2 dependency, only import it here
         try:
             import pandas as pd
         except ImportError as ex:
-            raise ImportError('Exporting to pandas needs a working installation of pandas. '
-                              'Importing pandas failed: ' + str(ex))
+            raise ImportError('Exporting to pandas needs a working installation'
+                              ' of pandas. Importing pandas failed: ' + str(ex))
         if units:
-            raise NotImplementedError('Units not supported when transforming to pandas')
-        # we take adventage of already implemented exporter
+            raise NotImplementedError('Units not supported when exporting to '
+                                      'pandas data frame')
+        # we take advantage of the already implemented exporter
         data = DictImportExport.export_data(group, variables,
                                             units=units, level=level)
         pandas_data = pd.DataFrame(data)
@@ -54,17 +59,19 @@ class PandasImportExport(ImportExport):
 
     @staticmethod
     def import_data(group, data, units=True, level=0):
-        # as pandas is not a default brian2 dependency we import it only in that namespace
+        # pandas is not a default brian2 dependency, only import it here
         try:
             import pandas as pd
         except ImportError as ex:
-            raise ImportError('Exporting to pandas needs a working installation of pandas. '
-                              'Importing pandas failed: ' + str(ex))
+            raise ImportError('Exporting to pandas needs a working installation'
+                              ' of pandas. Importing pandas failed: ' + str(ex))
         if units:
-            raise NotImplementedError('Units not supported when reading from pandas')
+            raise NotImplementedError('Units not supported when importing from '
+                                      'pandas data frame')
         colnames = data.columns
         array_data = data.values
         for e, colname in enumerate(colnames):
             if getattr(group.variables[colname], 'read_only'):
                 raise TypeError('Variable {} is read-only.'.format(colname))
-            group.state(colname, use_units=units, level=level+1)[:] = np.squeeze(array_data[:, e])
+            state = group.state(colname, use_units=units, level=level+1)
+            state[:] = np.squeeze(array_data[:, e])